This is the content xml for info page
This has the actual code for the imageView, but it's broken right now. 
To add an image right-click "res" > new > Image Asset
Then choose what image to load and it'll be created in the "mipmap" file.
This is clearly not how to load what we want but I'm using this method for now.